Improve the patient common lib structure
