+++ packages/esm-patient-common-lib/src/orders/index.ts
@@ -3,3 +3,5 @@ export * from './postOrders';
 export * from './useOrders';
 export * from './types';
 export * from './useOrderableConceptSets';
+export * from './useOrderType';
+export * from './useOrderTypes';
