+++ packages/esm-patient-common-lib/src/orders/types/index.ts
@@ -1,9 +1,4 @@
-import type { DrugOrderBasketItem } from './drug-order';
-import type { OrderBasketItem } from './order';
-import type { TestOrderBasketItem } from './test-order';
-
 export * from './order';
 export * from './drug-order';
 export * from './test-order';
-
-export type GenericOrderBasketItem = OrderBasketItem | DrugOrderBasketItem | TestOrderBasketItem;
+export * from './order-type';
