+++ packages/esm-patient-common-lib/src/orders/useOrderType.ts
@@ -1,0 +1,20 @@
+import { type FetchResponse, restBaseUrl } from '@openmrs/esm-framework';
+import { useMemo } from 'react';
+import useSWRImmutable from 'swr/immutable';
+import type { OrderTypeResponse } from './types';
+
+export function useOrderType(orderTypeUuid: string) {
+  const { data, isLoading, isValidating, error } = useSWRImmutable<FetchResponse<OrderTypeResponse>>(
+    `${restBaseUrl}/ordertype/${orderTypeUuid}`,
+  );
+  const results = useMemo(
+    () => ({
+      isLoadingOrderType: isLoading,
+      orderType: data?.data,
+      errorFetchingOrderType: error,
+      isValidatingOrderType: isValidating,
+    }),
+    [data?.data, error, isLoading, isValidating],
+  );
+  return results;
+}
