+++ packages/esm-patient-common-lib/src/orders/useOrderTypes.ts
@@ -1,0 +1,18 @@
+import { type FetchResponse, openmrsFetch, restBaseUrl } from '@openmrs/esm-framework';
+import useSWRImmutable from 'swr/immutable';
+import type { OrderTypeFetchResponse } from './types';
+
+export function useOrderTypes() {
+  const orderTypesUrl = `${restBaseUrl}/ordertype`;
+  const { data, error, isLoading, isValidating } = useSWRImmutable<FetchResponse<OrderTypeFetchResponse>, Error>(
+    orderTypesUrl,
+    openmrsFetch,
+  );
+
+  return {
+    data: data?.data?.results ?? null,
+    error,
+    isLoading,
+    isValidating,
+  };
+}
